Reading not working

+++ SharedMemory.cpp
@@ -2,4 +2,4 @@
 
 char * shmBlock::startPtr;
 char * shmBlock::lastUsed;
-size_t shmBlock::offset;
+int shmBlock::offset;
